Implement a trajectory_constraint `final_seed` for Pilz/LIN
The final seed can be set by the user via `trajectory_constraints` with the name `final_seed`.
This allows to augment the probability that the final point of the trajectory is close to the
required position in joint space.
This is especially useful for redundant robots that have an infinite number of solutions for a given Cartesian
goal.
With a joint-space goal, it is expected that the user provides a seed that is the same as the goal position
in joint space but this is not a requirement.
If a final seed is given, the seed will be interpolated between start- and end seeds rather than the classical
way, where the seed for an interpolated Cartesian pose is the result of the IK of the previous Cartesian pose in the trajectory.

There is not guarantee that the final joint pose will be the same
as the goal joint pose  but it is expected that it'll be closer than
when not using the final seed at all because, previously, the goal pose is
solely used to compute the FK before interpolating in Cartesian space.

Signed-off-by: Gaël Écorchard <gael.ecorchard@cvut.cz>